# openapi.yaml
openapi: 3.0.0
info:
  title: Bank API
  version: 1.0.0
  description: Bank API with interoperability features
  contact:
    name: API Support
    email: support@eerovallistu.site
servers:
  - url: /
    description: Current server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Transaction:
      type: object
      properties:
        id:
          type: integer
          description: Transaction ID
        fromAccount:
          type: string
          description: Source account number
        toAccount:
          type: string
          description: Destination account number
        amount:
          type: number
          format: float
          description: Transaction amount
        currency:
          type: string
          enum: [EUR, USD, GBP]
          description: Transaction currency
        explanation:
          type: string
          description: Transaction description
        senderName:
          type: string
          description: Name of the sender
        receiverName:
          type: string
          description: Name of the receiver
        status:
          type: string
          enum: [pending, inProgress, completed, failed]
          description: Transaction status
        isExternal:
          type: boolean
          description: Whether this is an external transaction
        createdAt:
          type: string
          format: date-time
          description: Transaction creation date
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                example: error
              message:
                type: string
                example: Authentication required
    ValidationError:
      description: Input validation error
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                example: error
              errors:
                type: array
                items:
                  type: object
                  properties:
                    field:
                      type: string
                    message:
                      type: string
security:
  - bearerAuth: []
paths:
...
  /transfers:
    get:
      summary: Get all transactions for the user
      tags: [Transactions]
      responses:
        '200':
          description: List of user transactions
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Transaction'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /transfers/internal:
    post:
      summary: Create a new internal transaction
      tags: [Transactions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - fromAccount
                - toAccount
                - amount
                - explanation
              properties:
                fromAccount:
                  type: string
                  example: 353c8b72e4a9f15d3b82
                toAccount:
                  type: string
                  example: 353f9a23d1c7b45e8t91
                amount:
                  type: number
                  example: 150.00
                explanation:
                  type: string
                  example: Rent payment for January
      responses:
        '201':
          description: Transaction created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    $ref: '#/components/schemas/Transaction'
        '400':
          $ref: '#/components/responses/ValidationError'
        '402':
          description: Insufficient funds
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: Insufficient funds
        '404':
          description: Account not found
  /transfers/external:
    post:
      summary: Create a new external transaction
      tags: [Transactions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - fromAccount
                - toAccount
                - amount
                - explanation
              properties:
                fromAccount:
                  type: string
                  example: 353c8b72e4a9f15d3b82
                toAccount:
                  type: string
                  example: 512f9a23d1c7b45e8t91
                amount:
                  type: number
                  example: 150.00
                explanation:
                  type: string
                  example: Payment for services
      responses:
        '201':
          description: Transaction created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    $ref: '#/components/schemas/Transaction'
        '400':
          $ref: '#/components/responses/ValidationError'
        '402':
          description: Insufficient funds
        '404':
          description: Account not found
  /transfers/{id}:
    get:
      summary: Get transaction details
      tags: [Transactions]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Transaction details
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    $ref: '#/components/schemas/Transaction'
        '404':
          description: Transaction not found
        '403':
          description: Access forbidden
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /transfers/history:
    get:
      summary: Get user's transaction history
      tags: [Transactions]
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, inProgress, completed, failed]
        - in: query
          name: type
          schema:
            type: string
            enum: [incoming, outgoing, all]
      responses:
        '200':
          description: Transaction history
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Transaction'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /transfers/incoming:
    post:
      summary: Process incoming transfer
      tags: [Bank-to-Bank]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - jwt
              properties:
                jwt:
                  type: string
                  example: eyJhbGciOiJSUzI1NiIsImtpZCI6IjEiLCJ0eXAiOiJKV1QifQ...
      responses:
        '200':
          description: Transaction processed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  receiverName:
                    type: string
                    example: John Smith
                  transactionId:
                    type: integer
                    example: 42
        '400':
          description: Invalid JWT or signature
        '404':
          description: Destination account not found
        '500':
          description: Processing error
...